// src/utils/enums.ts
export enum AppEvents {
	ItemsChanged = 'items:changed',
	ModalOpened = 'modal:opened',
	ModalClosed = 'modal:closed',
	PreviewChanged = 'preview:changed',
	CardSelected = 'card:selected',
	CardBasket = 'card:basket',
	OrderReady = 'order:ready',
	OrderOpened = 'order:opened',
	OrderSubmitted = 'order:submitted',
	OrderSuccess = 'order:success',
	AddressChanged = 'address:changed',
	AddressErrorsChanged = 'addressErrors:changed',
	ContactsErrorsChanged = 'contactsErrors:changed',
	ContactsChanged = 'contacts:changed',
	ContactsSubmitted = 'contacts:submitted',
	BasketChanged = 'basket:changed',
	BasketOpened = 'basket:opened',
	BasketAdded = 'basket:added',
	BasketRemoved = 'basket:removed',
}